requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
Appium-Python-Client==0.50
atomicwrites==1.3.0
attrs==19.3.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
facebook-wda==0.7.5
idna==2.9
importlib-metadata==1.6.0
inflect==4.1.0
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
poium==0.5.3
py==1.8.1
pycryptodome==3.9.7
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
pytest-assume==2.2.1
pytest-ordering==0.6
requests==2.23.0
retry==0.9.2
selenium==3.141.0
six==1.14.0
sqlacodegen==2.1.0
SQLAlchemy==1.3.16
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
